+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## Unreleased
 
+
+## 4.2.0
+
 * `laterpay.signing.sign_and_encode()` is deprecated and will be removed in a
   future release. Consider using `laterpay.utils.signed_query()` instead.
 * `signing.sign_get_url()` is deprecated and will be removed in a future